+++ tests/codegen/struct_nested_struct_named.cpp
@@ -28,7 +28,7 @@ entry:
   store i32 %3, i32* %2, align 8
   %4 = bitcast i32* %Bar.x to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %4)
-  %probe_read = call i64 inttoptr (i64 4 to i64 (i8*, i64, i8*)*)(i32* nonnull %Bar.x, i64 4, i8* nonnull %1)
+  %probe_read = call i64 inttoptr (i64 4 to i64 (i32*, i64, i8*)*)(i32* nonnull %Bar.x, i64 4, i8* nonnull %1)
   %5 = load i32, i32* %Bar.x, align 4
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %4)
   %6 = bitcast i64* %"@x_key" to i8*
@@ -39,7 +39,7 @@ entry:
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %8)
   store i64 %7, i64* %"@x_val", align 8
   %pseudo = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %update_elem = call i64 inttoptr (i64 2 to i64 (i8*, i8*, i8*, i64)*)(i64 %pseudo, i64* nonnull %"@x_key", i64* nonnull %"@x_val", i64 0)
+  %update_elem = call i64 inttoptr (i64 2 to i64 (i8*, i64*, i64*, i64)*)(i64 %pseudo, i64* nonnull %"@x_key", i64* nonnull %"@x_val", i64 0)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %6)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %8)
   ret i64 0
@@ -119,7 +119,7 @@ entry:
   %Bar.x = alloca i32, align 4
   %1 = bitcast i32* %Bar.x to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %1)
-  %probe_read = call i64 inttoptr (i64 4 to i64 (i8*, i64, i8*)*)(i32* nonnull %Bar.x, i64 4, i8* null)
+  %probe_read = call i64 inttoptr (i64 4 to i64 (i32*, i64, i8*)*)(i32* nonnull %Bar.x, i64 4, i8* null)
   %2 = load i32, i32* %Bar.x, align 4
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %1)
   %3 = bitcast i64* %"@x_key" to i8*
@@ -130,7 +130,7 @@ entry:
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %5)
   store i64 %4, i64* %"@x_val", align 8
   %pseudo = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %update_elem = call i64 inttoptr (i64 2 to i64 (i8*, i8*, i8*, i64)*)(i64 %pseudo, i64* nonnull %"@x_key", i64* nonnull %"@x_val", i64 0)
+  %update_elem = call i64 inttoptr (i64 2 to i64 (i8*, i64*, i64*, i64)*)(i64 %pseudo, i64* nonnull %"@x_key", i64* nonnull %"@x_val", i64 0)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %3)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %5)
   ret i64 0
